Moved helper class to service package
