+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/PdfCreationService.java
@@ -12,7 +12,7 @@ import org.thymeleaf.spring5.SpringTemplateEngine;
 import uk.gov.hmcts.reform.pip.publication.services.config.ThymeleafConfiguration;
 import uk.gov.hmcts.reform.pip.publication.services.models.external.Artefact;
 import uk.gov.hmcts.reform.pip.publication.services.service.pdf.converters.Converter;
-import uk.gov.hmcts.reform.pip.publication.services.utils.DateTimeHelper;
+import uk.gov.hmcts.reform.pip.publication.services.service.helpers.DateTimeHelper;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
