+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/service/helpers/DateTimeHelperTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pip.publication.services.utils;
+package uk.gov.hmcts.reform.pip.publication.services.service.helpers;
 
 import org.junit.jupiter.api.Test;
